restructure subject cards
